src/main/java/com/asteria/productcartservice/exception/ProductExceptionHandler.java
package com.asteria.productcartservice.exception;

import com.asteria.productcartservice.product.model.ProductErrorResponse;
import jakarta.persistence.EntityNotFoundException;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.ControllerAdvice;
import org.springframework.web.bind.annotation.ExceptionHandler;

@ControllerAdvice
public class ProductExceptionHandler {

    @ExceptionHandler(EntityNotFoundException.class)
    public ResponseEntity<ProductErrorResponse> handleProductNotFoundException(EntityNotFoundException ex) {
        ProductErrorResponse errorResponse = new ProductErrorResponse();
        errorResponse.setCode(HttpStatus.NOT_FOUND.value());
        errorResponse.setMessage(ex.getMessage());
        return new ResponseEntity<>(errorResponse, HttpStatus.NOT_FOUND);
    }

    @ExceptionHandler(IllegalArgumentException.class)
    public ResponseEntity<ProductErrorResponse> handleProductNotFoundException(IllegalArgumentException ex) {
        ProductErrorResponse errorResponse = new ProductErrorResponse();
        errorResponse.setCode(HttpStatus.NOT_FOUND.value());
        errorResponse.setMessage(ex.getMessage());
        return new ResponseEntity<>(errorResponse, HttpStatus.NOT_FOUND);
    }
}
